+++ src/main/java/grep/neogul_coder/domain/users/repository/UserRepository.java
@@ -1,6 +1,6 @@
-package grep.neogul_coder.users.repository;
+package grep.neogul_coder.domain.users.repository;
 
-import grep.neogul_coder.users.entity.User;
+import grep.neogul_coder.domain.users.entity.User;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
