+++ src/main/java/grep/neogul_coder/global/auth/service/AuthService.java
@@ -8,8 +8,8 @@ import grep.neogul_coder.global.auth.jwt.dto.TokenDto;
 import grep.neogul_coder.global.auth.oauth.user.OAuth2UserInfo;
 import grep.neogul_coder.global.auth.payload.LoginRequest;
 import grep.neogul_coder.global.auth.repository.UserBlackListRepository;
-import grep.neogul_coder.users.entity.User;
-import grep.neogul_coder.users.repository.UserRepository;
+import grep.neogul_coder.domain.users.entity.User;
+import grep.neogul_coder.domain.users.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
